pushing up utility files before branch switch

+++ LambdaZero/inputs/inputs_op.py
@@ -805,13 +805,16 @@ class BrutalDock(InMemoryDataset):
 
     def process(self):       
         print("processing", self.raw_paths)
+        
         for raw_path, processed_path in zip(self.raw_paths, self.processed_paths):
-            docked_index = pd.read_feather(raw_path)
-            smis = docked_index["smi"].tolist()
-            props = {pr:docked_index[pr].tolist() for pr in self._props}
-            tasks = [self.proc_func.remote(smis[j], {pr: props[pr][j] for pr in props},
+            if not os.path.isfile(processed_path):
+                docked_index = pd.read_feather(raw_path)
+                smis = docked_index["smi"].tolist()
+                props = {pr:docked_index[pr].tolist() for pr in self._props}
+                tasks = [self.proc_func.remote(smis[j], {pr: props[pr][j] for pr in props},
                                            self.pre_filter, self.pre_transform) for j in range(len(smis))]
-            graphs = ray.get(tasks)
-            graphs = [g for g in graphs if g is not None]
-            # save to the disk
-            torch.save(self.collate(graphs), processed_path)
+                graphs = ray.get(tasks)
+                graphs = [g for g in graphs if g is not None]
+                # save to the disk
+                torch.save(self.collate(graphs), processed_path)
+
